Moved Security and Localizer registration from core to runtime.
