+++ Source/Rhetos.Configuration.Autofac/Modules/RuntimeModule.cs
@@ -20,8 +20,10 @@
 using Autofac;
 using Rhetos.Dom;
 using Rhetos.Dsl;
+using Rhetos.Extensibility;
 using Rhetos.Persistence;
 using Rhetos.Processing;
+using Rhetos.Security;
 using Rhetos.Utilities;
 using Rhetos.XmlSerialization;
 
@@ -32,6 +34,9 @@ namespace Rhetos.Configuration.Autofac.Modules
         protected override void Load(ContainerBuilder builder)
         {
             var pluginRegistration = builder.GetPluginRegistration();
+
+            AddSecurity(builder, pluginRegistration);
+            AddUtilities(builder, pluginRegistration);
 
             builder.Register(context => context.Resolve<IConfigurationProvider>().GetOptions<RhetosAppOptions>()).SingleInstance().PreserveExistingDefaults();
             builder.RegisterType<DomLoader>().As<IDomainObjectModel>().SingleInstance();
@@ -47,6 +52,27 @@ namespace Rhetos.Configuration.Autofac.Modules
             pluginRegistration.FindAndRegisterPlugins<ICommandInfo>();
 
             base.Load(builder);
+        }
+
+        private void AddSecurity(ContainerBuilder builder, ContainerBuilderPluginRegistration pluginRegistration)
+        {
+            // TODO: SecurityOptions should probably not be required build container and possibly even for dbupgrade. Move to specific module registrations after refactor.
+            builder.Register(context => context.Resolve<IConfigurationProvider>().GetOptions<SecurityOptions>()).SingleInstance().PreserveExistingDefaults();
+            builder.RegisterType<WindowsSecurity>().As<IWindowsSecurity>().SingleInstance();
+            builder.RegisterType<AuthorizationManager>().As<IAuthorizationManager>().InstancePerLifetimeScope();
+
+            // Default user authentication and authorization components. Custom plugins may override it by registering their own interface implementations.
+            builder.RegisterType<NullAuthorizationProvider>().As<IAuthorizationProvider>().PreserveExistingDefaults();
+
+            // Cannot use FindAndRegisterPlugins on IUserInfo because each type should be manually registered with InstancePerLifetimeScope.
+            pluginRegistration.FindAndRegisterPlugins<IAuthorizationProvider>();
+            pluginRegistration.FindAndRegisterPlugins<IClaimProvider>();
+        }
+
+        private void AddUtilities(ContainerBuilder builder, ContainerBuilderPluginRegistration pluginRegistration)
+        {
+            pluginRegistration.FindAndRegisterPlugins<ILocalizer>();
+            builder.RegisterType<NoLocalizer>().As<ILocalizer>().SingleInstance().PreserveExistingDefaults();
         }
     }
 }
